# refactor: use unsafe to avoid alloc